Centralizing codec negotiation
This change places all codec negotiation into new libav.encoder and
libav.decoder functions, which handle generating a complete libav.js
codec name, context, and options dictionary. This will allow better
parsing and handling of codec details in the future.

Additionaly, this adds the ability to bypass the codec string entirely
and give a libav.js codec name, context, and options dictionary
directly. Obviously that will only work with the polyfill, but it allows
us to use codecs that are not in the registry.